Refactor role, rank checks
